src/auth_project/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-oauth-toolkit==1.1.2
djangorestframework==3.8.2
idna==2.6
oauth2-provider==0.0
oauthlib==2.0.7
pytz==2018.4
requests==2.18.4
urllib3==1.22
